Address Matt' comments
