--- pkg/controllers/vrpq/listrestorepoints_reconciler_test.go
@@ -1,74 +1,0 @@
-/*
- (c) Copyright [2021-2023] Open Text.
- Licensed under the Apache License, Version 2.0 (the "License");
- You may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
- http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-*/
-
-package vrpq
-
-import (
-	"context"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	v1 "github.com/vertica/vertica-kubernetes/api/v1"
-	vapi "github.com/vertica/vertica-kubernetes/api/v1beta1"
-	"github.com/vertica/vertica-kubernetes/pkg/cmds"
-	vmeta "github.com/vertica/vertica-kubernetes/pkg/meta"
-	"github.com/vertica/vertica-kubernetes/pkg/test"
-	ctrl "sigs.k8s.io/controller-runtime"
-)
-
-const TestPassword = "test-pw"
-
-var _ = Describe("stopdb_reconcile", func() {
-	ctx := context.Background()
-
-	It("should update query conditions with vclusterOps", func() {
-		vdb := v1.MakeVDB()
-		vdb.Annotations[vmeta.VClusterOpsAnnotation] = vmeta.VClusterOpsAnnotationTrue
-		test.CreateVDB(ctx, k8sClient, vdb)
-		defer test.DeleteVDB(ctx, k8sClient, vdb)
-
-		vrpq := vapi.MakeVrpq()
-		Expect(k8sClient.Create(ctx, vrpq)).Should(Succeed())
-		defer func() { Expect(k8sClient.Delete(ctx, vrpq)).Should(Succeed()) }()
-		fpr := &cmds.FakePodRunner{}
-		dispatcher := vrpqRec.makeDispatcher(logger, vdb, fpr, TestPassword)
-		recon := MakeRestorePointsQueryReconciler(vrpqRec, vrpq, logger, dispatcher)
-
-		Expect(recon.Reconcile(ctx, &ctrl.Request{})).Should(Equal(ctrl.Result{}))
-		// make sure that Quering condition is updated to Unknown and
-		// QueryComplete condition is updated to True
-		Expect(vrpq.IsStatusConditionTrue(vapi.Querying)).ShouldNot(BeTrue())
-		Expect(vrpq.IsStatusConditionTrue(vapi.QueryComplete)).Should(BeTrue())
-
-	})
-
-	It("should not update conditions with admintools", func() {
-		vdb := v1.MakeVDB()
-		test.CreateVDB(ctx, k8sClient, vdb)
-		defer test.DeleteVDB(ctx, k8sClient, vdb)
-
-		vrpq := vapi.MakeVrpq()
-		Expect(k8sClient.Create(ctx, vrpq)).Should(Succeed())
-		defer func() { Expect(k8sClient.Delete(ctx, vrpq)).Should(Succeed()) }()
-		fpr := &cmds.FakePodRunner{}
-		dispatcher := vrpqRec.makeDispatcher(logger, vdb, fpr, TestPassword)
-		recon := MakeRestorePointsQueryReconciler(vrpqRec, vrpq, logger, dispatcher)
-		result, err := recon.Reconcile(ctx, &ctrl.Request{})
-		Expect(result).Should(Equal(ctrl.Result{}))
-		Expect(err.Error()).To(ContainSubstring("ListRestorePoints is not supported in the admintools side"))
-		Expect(vrpq.IsStatusConditionTrue(vapi.QueryComplete)).ShouldNot(BeTrue())
-
-	})
-})
